Improved support for sync.(RW)Mutex and sync.WaitGroup
Both now have processes similar to the channel process, with a bad state.
Queries now check for the reachability of the bad state.
ir.SpecialOp and ir.SpecialOpStmt are used to track channel creation and
operations, and mutex and wait group creation and operations within the function
call graph.
A remaining TODO is to support defer for these operations.