Implemented panic and recover
Panic behaves like return and is integreated into ir.ReturnStmt. Recover is
represented with a new ir.RecoverStmt.

Each function in Uppaal has two bool variables relating to panic, one internal
and one external. An internal panic is a panic in the function itself or a
function it called. An external panic is a panic in the (direct) caller.
A recover statement sets the external panic to false.
An intenal panic gets passed to deferred functions as their external
panics. If no deferred functions recovers it, the internal panic becomes
external at the very end of the function.