Made "created" states in function calls committed
The created state gets reached after a new instance of a function process
has been set up for execution, including setting all of its arguments.
These steps happen as part of the update of the transition leading to the
"created" state and do not have any effect on the concurrency behaviour of
other processes (beyond possibly running out of resources earlier).
Therefore it makes sense to make the "created" state committed to force
the following regular, or, defer call to happen immediately. This is
always possible.