Added arrays, slices, and maps
Arrays, slices, and maps are collectively referred to as container types and
represented by ir.ContainerType. The ir, builder, analyzer, and translator have
been modified similarly to the implementation of structs. Read and write
accesses (which are differentiated) are represented by ir.ContainerAccess, which
also holds index information.
ir.AppendIndex is used to signify a slice append operation. append() currently
gets modeled as first a slice copy, followed by appending to the copied slice
in place. In the future this can be optimized for re-assignments, e.g.

x = append(x, y)

ir.RandomIndex is used to signify a random container access. Uppaal select
statements are used to generated random indices. For arrays, the random index
is in the range of statically determinable valid array indices, since array
size is always known at compile time. For slices, the random index is in the
range of possible slice slots (bound by the container capacity config value).
Guards are used to ensure that random indices that exceed the current slice
length are not used.
Map accesses with a random index are treated differently for reads and writes.
For reads, the index is between -1 and the set container capacity and a guard
ensures it is less than the map length. The index -1 models a key not found in
the map, resulting in the default value being returned (also see below). For
writes, the index is between 0 and the set container capacity (inclusive) and a
guard ensures it is less than or equal to the map length. An index equal to the
map length is interpreted as inserting a new value in the map. A lower index
corresponds to overwriting a map value.
The translator has a new randomVariableSupplier (rvs) to help with the random
variable generation in select statements, generation of appropriate transition
guards.

Range loops over containers are supported. The loop variables for the current
iteration are provided, if possible (also for regular for loops).

Some other general changes were also necessary:

- Values for statements that require multiple transitions, e.g. sending on a
  channel or a range loop, are evaluated only once in the first transition now.
  This became necessary because random container accesses should only happen
  once. However, this also fixes an issue where it would have been possible to
  change the used value mid-operation, e.g. send via the trigger channel and
  receive via the confirm channel of two different Channel processes.
- There are new special ir.Values to signify an initialized mutex, wait group,
  struct, or array. The translator translates these into calls to the
  respective make function in Uppaal. Some make functions now also have a bool
  argument to signify if the members should be recursively initialized.
  This made it possible to remove recursive initialization from the builder.
  The builder also no longer provides initial values, this functionality has
  been moved to the ir. The analyzer takes this into account to determine
  resource bounds.
  This change became necessary to enable returning a default map value for map
  read operations that access a missing map key.
- Variadic functions are now supported. Trailing arguments get combined into a
  slice variable, similar to slice literals.
- Initialization functions in declarations are now generated last because they
  need to refer to other prior functions, such as make functions for structs
  and arrays, which were previously handled in explicit ir make statements for
  variable initialization.