Added analyzer for type and variable use and additional optimizations
The new analyzer.VarInfo holds information on which variables and functions
use types and which functions use variables. This is used to skip declaring
types and variables in the translator if IR optimizations are turned on and
the types and variables in question are unused. Types are used if they are
either unused or only used in variables and functions that are unused.
Variables are unused if they are unused or only used in unused functions.
For most test cases this mostly only excludes Int, Func, Mutex, and WaitGroup.
However, for test entry functions in bolt this is also removing some variables
and struct types.